カレンダーのEntityを追加
