+++ ios/Parjob/Parjob/Calendar/CalendarViewModel.swift
@@ -16,44 +16,6 @@ protocol CalendarModelDelegate: class {
     func faildAPI(title: String, msg: String)
 }
 
-struct TableViewShift {
-    var joined: String = ""
-    var shifts: [UserShift] = []
-    
-    mutating func generateJoinedString() {
-        var tmp = ""
-        self.shifts.forEach { (userShift) in
-            tmp += userShift.user + " "
-        }
-        
-        joined = tmp.substring(to: tmp.index(before: tmp.endIndex))
-    }
-}
-
-struct UserShift {
-    var id: Int = 0
-    var name: String = ""
-    var user: String = ""
-}
-
-struct ShiftCategory {
-    var name: String = ""
-    var userShift: [UserShift] = []
-}
-
-struct TargetUserShift {
-    var id: Int = 0
-    var name: String = ""
-    var color: String = ""
-}
-
-struct OneDayShift {
-    var date: String = ""
-    var memo: String = ""
-    var shift: [ShiftCategory] = []
-    var user: TargetUserShift = TargetUserShift()
-}
-
 class CalendarModel {
     weak var delegate: CalendarModelDelegate?
     private let api = API()
